# Conflicts:
#	Programs/ResourceEditor/Classes/CommandLine/Private/Dump/DumpTool.cpp


Former-commit-id: af4f2dc4c9d80d53060eee59d85f9a709b1c2973

+++ Sources/Internal/Render/RHI/DX9/rhi_TextureDX9.cpp
@@ -83,11 +83,11 @@ bool TextureDX9_t::Create(const Texture::Descriptor& desc, bool forceExecute)
 
         if (is_depthbuf)
         {
-            cmds[0] = { DX9Command::CREARE_DEPTHSTENCIL_SURFACE, { desc.width, desc.height, fmt, desc.sampleCount, 0, 0, uint64_t(&surf9), 0 } };
+            cmds[0] = { DX9Command::CREARE_DEPTHSTENCIL_SURFACE, { desc.width, desc.height, uint64(fmt), desc.sampleCount, 0, 0, uint64_t(&surf9), 0 } };
         }
         else
         {
-            cmds[0] = { DX9Command::CREATE_RENDER_TARGET, { desc.width, desc.height, fmt, desc.sampleCount, 0, 0, uint64_t(&surf9), 0 } };
+            cmds[0] = { DX9Command::CREATE_RENDER_TARGET, { desc.width, desc.height, uint64(fmt), desc.sampleCount, 0, 0, uint64_t(&surf9), 0 } };
         };
         ExecDX9(cmds, countof(cmds), forceExecute);
         DVASSERT(surf9 != nullptr);
